dto: Remove cycles in wiring DTOs
The set of referenced DTOs is pushed up to the top level BundleWiringDTO
and each DTO is assigned a transient identifier that can be used as a
reference by other DTOs. An example transient identifier could be the
identity hash code of the underlying runtime object.

Since CapabilityDTO and RequirementDTO object can be large, due to the
contained directives and attributes maps, we change to use ref DTOs in
the wiring DTO types. The ref DTOs refer to the CapabilityDTO/
RequirementDTO objects in the BundleRevisionDTO.